refactor(mining): one file per probe strategy + collapse/context
Pure structural move (no behavior change): parameter_analysis/mining.rs
(2066 lines) is now parameter_analysis/mining/ —
- mod.rs             module docs + imports + mine_parameters orchestrator
- collapse.rs        EWMA collapse detection + slot snapshot/any-param folding
- context_detect.rs  the detect_injection_context / detect_js_breakout family
- probe_dictionary / probe_body / probe_response_id / probe_json /
  probe_graphql / probe_xml / probe_multipart — one probe strategy each

The cross-module detectors are re-exported (`pub(crate) use context_detect::*`)
so `parameter_analysis::mining::detect_injection_context_with_marker`, the
parent's `pub use mining::*`, and `pub(crate) use mining::detect_injection_context`
all resolve unchanged. Submodules share helpers/imports via `use super::*`;
tests.rs untouched (collapse-state fields/methods + a couple of consts are
pub(super) so the sibling test module still reaches them). 185
parameter_analysis tests pass, clippy clean.
