--- src/parameter_analysis/mining.rs
@@ -1,2066 +1,0 @@
-//! # Stage 2: Mining
-//!
-//! Discovers additional parameters by analyzing HTML forms, JavaScript source,
-//! dictionary wordlists, and GF-pattern lists — then probes each for reflection.
-//!
-//! **Input:** `Target` + `ScanArgs` + the initial `reflection_params` from Stage 1.
-//!
-//! **Output:** Extends the shared `reflection_params` list with newly discovered
-//! `Param` entries that reflect. Each carries the same naive `valid_specials`,
-//! `invalid_specials`, and `injection_context` as Stage 1 output.
-//!
-//! **Side effects:** HTTP requests for DOM/dict/GF mining probes. Uses EWMA-based
-//! collapse detection to short-circuit when a target reflects everything
-//! (sustained ≥85% reflection rate after ≥15 attempts). Filters out 5xx
-//! responses to avoid false positives from debug/error pages.
-//!
-//! **Skippable via:** `--skip-mining`, `--skip-mining-dict`, `--skip-mining-dom`.
-
-use crate::cmd::scan::ScanArgs;
-use crate::parameter_analysis::{DelimiterType, InjectionContext, Location, Param};
-use crate::payload::mining::GF_PATTERNS_PARAMS;
-use crate::target_parser::Target;
-use crate::utils::shimmer::ShimmerSpinner;
-use std::sync::Arc;
-
-use tokio::sync::{Mutex, Semaphore};
-use tokio::time::{Duration, sleep};
-use url::form_urlencoded;
-
-use crate::scanning::selectors;
-
-const EWMA_ALPHA: f64 = 0.30; // smoothing factor for exponential weighted moving average
-const EWMA_START_VALUE: f64 = 0.0;
-const COLLAPSE_EWMA_THRESHOLD: f64 = 0.85; // if sustained EWMA reflection ratio >= 85%
-const COLLAPSE_MIN_ATTEMPTS: usize = 15; // need at least this many attempts before collapsing
-const COLLAPSE_MIN_REFLECTIONS: usize = 5; // and at least this many reflections
-
-/// Number of random sentinel param names probed up-front to detect
-/// "reflect-everything" pages. Three is enough to make the false-positive
-/// rate negligible while keeping the ceiling cost low (≤ 3 wasted requests).
-const SENTINEL_PROBE_COUNT: usize = 3;
-
-/// Safety cap on DOM-extracted candidate parameter names probed by
-/// [`probe_response_id_params`]. A hostile/huge response body (up to the 16 MiB
-/// `read_body` cap) packed with distinct `<input id=…>` / `name=…` attributes
-/// can yield ~10^6 unique candidates; without a cap each became a spawned probe
-/// task (and an outbound request), scaling memory and request fan-out with the
-/// attacker-controlled body. Real pages have at most a few hundred distinct
-/// field names, so this ceiling is generous; truncation is surfaced as a warning.
-const MAX_DOM_MINING_PARAMS: usize = 4096;
-
-/// Cap a DOM-extracted candidate-name set to [`MAX_DOM_MINING_PARAMS`]. Returns
-/// the (possibly truncated) names plus `Some(original_len)` when truncation
-/// occurred (so the caller can warn), or `None` when already under the cap.
-/// Extracted so the boundary (`len == cap` not truncated, `len == cap + 1`
-/// truncated) is unit-testable without standing up a giant HTML body.
-fn cap_dom_params(mut names: Vec<String>) -> (Vec<String>, Option<usize>) {
-    let original = names.len();
-    if original > MAX_DOM_MINING_PARAMS {
-        names.truncate(MAX_DOM_MINING_PARAMS);
-        (names, Some(original))
-    } else {
-        (names, None)
-    }
-}
-
-/// Sentinel parameter names — random-looking, namespace-prefixed strings
-/// that should never collide with real params on a normal application.
-/// If every one of these reflects, the page is echoing arbitrary input
-/// and there's no point iterating a wordlist.
-const SENTINEL_QUERY_NAMES: &[&str] = &[
-    "dlfx_sentinel_q_8a3f",
-    "dlfx_canary_b27z_p1",
-    "dlfx_probe_xx9k_z2",
-];
-
-/// Run sentinel probes against the target by injecting the marker into a
-/// query parameter named `name` and checking whether the response body
-/// (or redirect Location header) echoes the marker. Returns the first
-/// response body when every sentinel reflects, or `None` as soon as any
-/// sentinel fails to reflect — `None` is the "this page is fine, run the
-/// wordlist normally" signal.
-async fn pre_collapse_query_probe(client: &reqwest::Client, target: &Target) -> Option<String> {
-    let marker = crate::scanning::markers::bracketed_marker();
-    let mut first_text: Option<String> = None;
-    for name in SENTINEL_QUERY_NAMES.iter().take(SENTINEL_PROBE_COUNT) {
-        let mut url = target.url.clone();
-        url.query_pairs_mut().append_pair(name, marker);
-        let req = crate::utils::build_request(
-            client,
-            target,
-            target.parse_method(),
-            url,
-            target.data.clone(),
-        );
-        crate::record_outbound_request().await;
-        let resp = crate::utils::http::send_counted(req).await.ok()?;
-        let location_has_marker = resp.status().is_redirection()
-            && resp
-                .headers()
-                .get("location")
-                .and_then(|v| v.to_str().ok())
-                .is_some_and(|loc| loc.contains(marker));
-        let text = crate::utils::http::read_body(resp).await.ok()?;
-        if !location_has_marker && !text.contains(marker) {
-            return None;
-        }
-        if first_text.is_none() {
-            first_text = Some(text);
-        }
-    }
-    first_text
-}
-
-/// Identity of a parameter *slot*, used to tell the params a mining stage
-/// inherited apart from the ones it mined itself. `Location` is not `Hash`, so
-/// the key is built from its `Debug` rendering — the same shape
-/// `discovery::dedupe_reflection_params` uses.
-fn param_slot_key(p: &Param) -> String {
-    format!("{}|{:?}", p.name, p.location)
-}
-
-/// Snapshot the parameter slots already present before a mining stage starts.
-///
-/// Mining stages run sequentially (see [`mine_parameters`]), so this is a
-/// stable "everything Stage 1 discovery — or an earlier mining stage — already
-/// found" set, and a later collapse can subtract from it safely.
-async fn snapshot_param_slots(
-    reflection_params: &Arc<Mutex<Vec<Param>>>,
-) -> std::collections::HashSet<String> {
-    reflection_params
-        .lock()
-        .await
-        .iter()
-        .map(param_slot_key)
-        .collect()
-}
-
-/// Build the synthetic `any` param standing in for "this target echoes
-/// arbitrary parameter names at `location`".
-///
-/// Metadata comes from whichever sample the caller has: the response body that
-/// proved arbitrary names reflect (sentinel path), or one of the params the
-/// stage mined before collapsing (EWMA path). Both describe how an *arbitrary*
-/// name behaves here, which is what `any` represents. The params the target
-/// actually carries are deliberately not used as the sample — their measured
-/// specials belong to them, and copying those onto `any` produced findings that
-/// reported one parameter's evidence under another's name.
-fn make_any_param(
-    location: Location,
-    response_text: Option<&str>,
-    mined_sample: Option<&Param>,
-) -> Param {
-    let mut param = Param {
-        injection_context: Some(InjectionContext::Html(None)),
-        ..Param::new(
-            "any".to_string(),
-            crate::scanning::markers::bracketed_marker().to_string(),
-            location,
-        )
-    };
-    if let Some(text) = response_text {
-        param = param.with_reflection_analysis(text);
-    } else if let Some(sample) = mined_sample {
-        param.value = sample.value.clone();
-        param.injection_context = sample.injection_context.clone();
-        param.valid_specials = sample.valid_specials.clone();
-        param.invalid_specials = sample.invalid_specials.clone();
-        param.js_breakout = sample.js_breakout.clone();
-    }
-    param
-}
-
-/// Collapse the params *this mining stage itself discovered* at `location`
-/// into the single synthetic `any` param, leaving everything else untouched.
-///
-/// Collapse is a cost control. Once a target is known to echo arbitrary
-/// parameter names, walking the rest of the wordlist only manufactures
-/// injection points that all behave identically, and each one costs a full
-/// Stage 3-6 payload run. Dropping *those* is the point of the mechanism.
-///
-/// Dropping the parameters the target actually carries is not. Those come from
-/// Stage 1 discovery — the URL query string, form fields, headers — and were
-/// confirmed to reflect before mining ever ran. Earlier revisions cleared every
-/// param at `location` regardless of origin, so on any page that echoes its
-/// whole query string (a debug endpoint, a framework error page that dumps
-/// `request.query_params`) the genuinely vulnerable parameter was deleted
-/// before Stage 3 ever saw it: the scan then reported a POC against `any` — a
-/// parameter the application does not have — and missed the real one entirely.
-/// `preexisting` is the guard against that, and it is why every caller must
-/// snapshot with [`snapshot_param_slots`] before it starts pushing.
-async fn collapse_mined_params(
-    reflection_params: &Arc<Mutex<Vec<Param>>>,
-    preexisting: &std::collections::HashSet<String>,
-    location: Location,
-    response_text: Option<&str>,
-) {
-    let mut guard = reflection_params.lock().await;
-    let mut mined_sample: Option<Param> = None;
-    guard.retain(|p| {
-        if p.location != location || preexisting.contains(&param_slot_key(p)) {
-            return true;
-        }
-        if mined_sample.is_none() {
-            mined_sample = Some(p.clone());
-        }
-        false
-    });
-    // Don't add a second `any` to a slot that already has one. That covers the
-    // synthetic left by an earlier stage collapsing the same location, and —
-    // more importantly — a target that genuinely carries a parameter called
-    // `any`: it is a real, discovered param and overwriting it with the
-    // synthetic stand-in would reintroduce exactly the data loss this function
-    // exists to prevent.
-    if !guard
-        .iter()
-        .any(|p| p.location == location && p.name == "any")
-    {
-        guard.push(make_any_param(
-            location,
-            response_text,
-            mined_sample.as_ref(),
-        ));
-    }
-}
-
-#[derive(Debug)]
-struct MiningSampleStats {
-    attempts: usize,
-    reflections: usize,
-    ewma_ratio: f64,
-    collapsed: bool,
-}
-
-impl MiningSampleStats {
-    fn new() -> Self {
-        Self {
-            attempts: 0,
-            reflections: 0,
-            ewma_ratio: EWMA_START_VALUE,
-            collapsed: false,
-        }
-    }
-    fn record_attempt(&mut self) {
-        self.attempts += 1;
-    }
-    fn record_reflection(&mut self) {
-        self.reflections += 1;
-        let instant = 1.0; // this attempt reflected
-        self.update_ewma(instant);
-    }
-    fn record_non_reflection(&mut self) {
-        let instant = 0.0;
-        self.update_ewma(instant);
-    }
-    fn update_ewma(&mut self, instant: f64) {
-        self.ewma_ratio = EWMA_ALPHA * instant + (1.0 - EWMA_ALPHA) * self.ewma_ratio;
-    }
-    fn should_collapse(&self) -> bool {
-        !self.collapsed
-            && self.attempts >= COLLAPSE_MIN_ATTEMPTS
-            && self.reflections >= COLLAPSE_MIN_REFLECTIONS
-            && self.ewma_ratio >= COLLAPSE_EWMA_THRESHOLD
-    }
-}
-
-pub(crate) fn detect_injection_context(text: &str) -> InjectionContext {
-    // Inner marker survives every reflection form classified by
-    // `classify_probe_reflection` (Full / PrefixOnly / SuffixOnly /
-    // InnerOnly), so it's the most reliable anchor for context inference
-    // on bracketed probes. Fall back to the open marker for callers that
-    // still inject it directly (older tests, legacy probe sites).
-    let inner = crate::scanning::markers::inner_marker();
-    if text.contains(inner) {
-        return detect_injection_context_with_marker(text, inner);
-    }
-    let open = crate::scanning::markers::open_marker();
-    detect_injection_context_with_marker(text, open)
-}
-
-/// Compute the exact JS breakout closer (issue #1073 follow-up) from the
-/// *observed* response when a probe marker reflects inside an inline `<script>`
-/// body. Returns the minimal closer sequence — produced by
-/// [`crate::payload::js_breakout::compute_js_breakout`] over the real script
-/// source from the enclosing `<script>` content start up to the reflection
-/// point — that escapes the open string and every unbalanced `([{` so a
-/// following `;<payload>//` reaches executable statement position.
-///
-/// This is the per-parameter carrier the synthesis layer consumes to emit a
-/// breakout matched to the *site's actual nesting* rather than only the fixed
-/// depth-0–3 catalog. Returns `None` when the marker is not inside an inline
-/// `<script>` body (the script-tag requirement scopes this to script contexts,
-/// not event-handler/attribute JS), when no marker is present, or when the
-/// observed prefix already sits at statement position (empty closer) — every
-/// such case falls back to the fixed catalog, so the result is strictly
-/// additive and never removes coverage.
-///
-/// Mirrors `detect_injection_context`'s marker selection (inner marker
-/// preferred, open marker fallback) so the closer is computed for the same
-/// reflection the context classifier anchored on. Uses raw response slicing
-/// (inline `<script>` content is CDATA-like, not HTML-entity-decoded by the
-/// browser), so the prefix matches the JS source the browser actually parses.
-pub(crate) fn detect_js_breakout(text: &str) -> Option<String> {
-    let inner = crate::scanning::markers::inner_marker();
-    let marker = if text.contains(inner) {
-        inner
-    } else {
-        crate::scanning::markers::open_marker()
-    };
-    detect_js_breakout_with_marker(text, marker)
-}
-
-/// `detect_js_breakout` with a caller-supplied marker string (mirrors
-/// `detect_injection_context_with_marker`). Used by probes that inject a
-/// non-standard marker, e.g. the numeric-only discovery probe.
-/// Offset of the last ASCII-case-insensitive occurrence of `needle` in
-/// `bytes[..before]`.
-///
-/// Allocation-free on purpose. The obvious spelling — `text[..mp]
-/// .to_ascii_lowercase().rfind(..)` — copies the whole prefix, and that prefix
-/// can be the full `read_body_capped` 16 MiB, once per probed parameter.
-/// Bounding the search window instead would have cost recall: framework apps
-/// routinely inline a serialized state blob or a whole bundle in one
-/// `<script>`, and a reflection several hundred KiB into it is ordinary, so any
-/// window short enough to be a useful cost bound also silently dropped those
-/// reflections back to the fixed catalog. Not allocating removes the need to
-/// choose. (`MAX_OPEN_DEPTH` in `js_breakout` remains the guard that stops a
-/// hostile prefix from producing a giant closer.)
-fn rfind_ascii_case_insensitive(bytes: &[u8], before: usize, needle: &[u8]) -> Option<usize> {
-    let before = before.min(bytes.len());
-    if needle.is_empty() || needle.len() > before {
-        return None;
-    }
-    (0..=before - needle.len())
-        .rev()
-        .find(|&start| bytes[start..start + needle.len()].eq_ignore_ascii_case(needle))
-}
-
-pub(crate) fn detect_js_breakout_with_marker(text: &str, marker: &str) -> Option<String> {
-    let mp = text.find(marker)?;
-    // Find the enclosing inline `<script …>` opening tag before the reflection.
-    // `<script` (case-insensitive) never matches a closing `</script>` tag (the
-    // char after `<` is `/`, not `s`), so the last match is a real opener.
-    let open_tag = rfind_ascii_case_insensitive(text.as_bytes(), mp, b"<script")?;
-    // Script content begins just after the `>` that ends the opening tag.
-    let gt = text[open_tag..mp].find('>')?;
-    let content_start = open_tag + gt + 1;
-    let prefix = &text[content_start..mp];
-    // Guard the multi-`<script>` edge: if a `</script>` closes between this
-    // opener and the marker, the marker isn't inside this script body — bail to
-    // the fixed catalog rather than computing a bogus closer.
-    if prefix.to_ascii_lowercase().contains("</script") {
-        return None;
-    }
-    let closer = crate::payload::js_breakout::compute_js_breakout(prefix);
-    // An empty closer means the prefix already sits at statement position; the
-    // raw-JS catalog templates handle that, so carry nothing.
-    if closer.is_empty() {
-        None
-    } else {
-        Some(closer)
-    }
-}
-
-/// Like `detect_injection_context` but uses a caller-supplied marker string.
-/// Useful for probes that don't use the standard alphanumeric marker (e.g. numeric-only probes).
-pub(crate) fn detect_injection_context_with_marker(text: &str, marker: &str) -> InjectionContext {
-    if !text.contains(marker) {
-        return InjectionContext::Html(None);
-    }
-
-    // Fast comment check using raw HTML when available
-    if let (Some(cs), Some(ce)) = (text.find("<!--"), text.find("-->"))
-        && let Some(mp) = text.find(marker)
-        && cs < mp
-        && mp < ce
-    {
-        return InjectionContext::Html(Some(DelimiterType::Comment));
-    }
-
-    // Parse HTML and locate marker via element text/attributes/script
-    let document = crate::utils::html::parse_document_bounded(text);
-
-    // Heuristic to infer surrounding quote delimiter around the first marker.
-    // Picks the *closest* opening quote before the marker (the one that
-    // actually contains it). Includes backtick template literals so a marker
-    // reflected inside `` `…` `` is reported as Backtick rather than falling
-    // back to None — the breakout payload (`${…}`) is different from `'/`"`.
-    fn infer_quote_delimiter(text: &str, marker: &str) -> Option<DelimiterType> {
-        let pos = text.find(marker)?;
-        let before = &text[..pos];
-        let after = &text[pos + marker.len()..];
-
-        let candidates: [(char, DelimiterType); 3] = [
-            ('"', DelimiterType::DoubleQuote),
-            ('\'', DelimiterType::SingleQuote),
-            ('`', DelimiterType::Backtick),
-        ];
-
-        let (qch, delim) = candidates
-            .iter()
-            .filter_map(|(c, d)| before.rfind(*c).map(|p| (*c, d.clone(), p)))
-            .max_by_key(|(_, _, p)| *p)
-            .map(|(c, d, _)| (c, d))?;
-
-        if after.find(qch).is_some() {
-            return Some(delim);
-        }
-        None
-    }
-
-    fn is_url_like_attribute(name: &str) -> bool {
-        matches!(
-            name.to_ascii_lowercase().as_str(),
-            "src" | "href" | "xlink:href" | "data" | "action" | "formaction" | "poster"
-        )
-    }
-
-    // 1) JavaScript context: marker appears in any <script> text
-    {
-        let sel = selectors::script();
-        for el in document.select(sel) {
-            let s = el.text().fold(String::new(), |mut acc, t| {
-                acc.push_str(t);
-                acc
-            });
-            if s.contains(marker) {
-                let delim = infer_quote_delimiter(text, marker);
-                return InjectionContext::Javascript(delim);
-            }
-        }
-    }
-
-    // 1b) CSS context: marker appears in any <style> text
-    {
-        let sel = selectors::style();
-        for el in document.select(sel) {
-            let s = el.text().fold(String::new(), |mut acc, t| {
-                acc.push_str(t);
-                acc
-            });
-            if s.contains(marker) {
-                let delim = infer_quote_delimiter(text, marker);
-                return InjectionContext::Css(delim);
-            }
-        }
-    }
-
-    // 2) Attribute context: marker in any attribute value.
-    //
-    // Event-handler attributes (`onload`, `onerror`, `onclick`, …) hold
-    // JavaScript source that the browser feeds into an event handler
-    // function — escaping out of the surrounding string literal yields
-    // a JS-context XSS, not an HTML one. Classifying these as plain
-    // `Attribute(delim)` makes the payload generator emit HTML tag
-    // breakouts (`'><svg…>`) that get serialised as inert HTML inside
-    // the handler's string. JS-breakout payloads (`',alert(1),'`) are
-    // the right strategy, so route on*-attributes through the
-    // `Javascript(delim)` branch.
-    fn is_event_handler_attribute(name: &str) -> bool {
-        let n = name.to_ascii_lowercase();
-        n.starts_with("on") && n.len() > 2
-    }
-    {
-        let any = selectors::universal();
-        for el in document.select(any) {
-            for (name, v) in el.value().attrs() {
-                if v.contains(marker) {
-                    let delim = infer_quote_delimiter(text, marker);
-                    if is_event_handler_attribute(name) {
-                        return InjectionContext::Javascript(delim);
-                    }
-                    return if is_url_like_attribute(name) {
-                        InjectionContext::AttributeUrl(delim)
-                    } else {
-                        InjectionContext::Attribute(delim)
-                    };
-                }
-            }
-            // Marker landed *as* an attribute name (not a value). Example:
-            //   <div id='x' MARKER>
-            // Scraper parses MARKER as a boolean attribute with empty value,
-            // so the value-side scan above misses it. This is the "free
-            // attribute slot inside an existing tag" position — HTML-tag
-            // breakouts (`<svg…>`) just become more attribute names, but
-            // bare event handlers (`onmouseover=alert(1)`) execute as-is.
-            // Classify as Attribute(None) so the payload generator emits
-            // the unquoted-attribute branch (event handlers + protocols).
-            for (name, _v) in el.value().attrs() {
-                if name.contains(marker) {
-                    return InjectionContext::Attribute(None);
-                }
-            }
-        }
-    }
-
-    // 3) HTML text context — deliberately not scanned.
-    //
-    // There used to be a loop here that walked every element and materialized
-    // `el.text()` (that element's *entire subtree* text) looking for the
-    // marker, returning `InjectionContext::Html(None)` on a hit. It could not
-    // change the answer: the fallback below is the same value, so the loop's
-    // only observable effect was its cost — and that cost is quadratic, since
-    // each of N elements re-collects the text of everything beneath it. A body
-    // of `<div>` × 500 000 with the marker in a *comment* node (which matches
-    // no element, so the loop never short-circuits) ran it to completion.
-    //
-    // Fallback to HTML.
-    InjectionContext::Html(None)
-}
-
-/// Identify a framework innerHTML-style sink the marker landed inside,
-/// if any. Returns the directive/attribute name (`"v-html"`,
-/// `"data-bind"`, `"ng-bind-html"`, …) so scanning can:
-///   1. Upgrade the finding's `inject_type` to surface the sink class.
-///   2. Treat HTML-entity-encoded reflections as exploitable — the
-///      framework hands the entity-decoded value to `innerHTML` at
-///      runtime, so `&lt;img onerror=…&gt;` still executes.
-///
-/// Conservative: only returns `Some(_)` when *every* marker occurrence
-/// sits inside one of the recognised attributes. A single occurrence in
-/// plain text content is enough to fall back to generic HTML payloads —
-/// the regular `detect_injection_context` already covers that path.
-///
-/// Returns `None` when:
-///   * the marker isn't present at all,
-///   * any occurrence lives outside an HTML attribute (text node,
-///     `<script>`, `<style>`), or
-///   * the attribute name isn't in the recognised innerHTML-sink set.
-pub(crate) fn detect_framework_html_sink(text: &str, marker: &str) -> Option<&'static str> {
-    if marker.is_empty() || !text.contains(marker) {
-        return None;
-    }
-    let document = crate::utils::html::parse_document_bounded(text);
-    let any = selectors::universal();
-    let mut found: Option<&'static str> = None;
-    for el in document.select(any) {
-        for (name, value) in el.value().attrs() {
-            if !value.contains(marker) {
-                continue;
-            }
-            let sink = match name.to_ascii_lowercase().as_str() {
-                "v-html" => Some("v-html"),
-                "ng-bind-html" | "[innerhtml]" | "innerhtml" => Some("ng-bind-html"),
-                // Knockout `data-bind` carries multiple clauses
-                // (e.g. `data-bind="text: foo, html: bar"`). Only the
-                // `html:` clause is an innerHTML sink, so require the
-                // clause to live at a real binding boundary — start of
-                // the attribute or after `,` / `;` / whitespace. A bare
-                // `value.contains("html:")` false-positives on
-                // `data-bind="text: 'html: link'"` and on any string
-                // literal that happens to contain `html:`.
-                "data-bind" if has_knockout_html_clause(value) => Some("data-bind"),
-                _ => None,
-            };
-            let s = sink?;
-            match found {
-                Some(prev) if prev != s => return None,
-                _ => found = Some(s),
-            }
-        }
-    }
-    found
-}
-
-/// True when `value` (a Knockout `data-bind` attribute) has an `html:`
-/// clause at a real binding boundary — start of the value, or after
-/// `,` / `;` / whitespace that separates clauses. Skipping inside
-/// quoted strings prevents the dominant false-positive shape:
-/// `data-bind="text: 'html: link'"` where `html:` is just data.
-fn has_knockout_html_clause(value: &str) -> bool {
-    let bytes = value.as_bytes();
-    let mut i = 0;
-    let mut quote: Option<u8> = None;
-    let mut at_clause_start = true;
-    while i < bytes.len() {
-        let b = bytes[i];
-        // Track string literal context so an `html:` substring inside a
-        // quoted clause value doesn't trigger.
-        if let Some(q) = quote {
-            if b == q {
-                quote = None;
-            }
-            i += 1;
-            continue;
-        }
-        if b == b'"' || b == b'\'' {
-            quote = Some(b);
-            at_clause_start = false;
-            i += 1;
-            continue;
-        }
-        if b == b',' || b == b';' {
-            at_clause_start = true;
-            i += 1;
-            continue;
-        }
-        if b.is_ascii_whitespace() {
-            i += 1;
-            continue;
-        }
-        if at_clause_start {
-            // Check for `html` followed by whitespace + `:`. ASCII-only
-            // attribute name, so direct byte comparison is fine.
-            let remaining = &bytes[i..];
-            if remaining.len() >= 4 && remaining[..4].eq_ignore_ascii_case(b"html") {
-                let mut j = i + 4;
-                while j < bytes.len() && bytes[j].is_ascii_whitespace() {
-                    j += 1;
-                }
-                if j < bytes.len() && bytes[j] == b':' {
-                    return true;
-                }
-            }
-            at_clause_start = false;
-        }
-        i += 1;
-    }
-    false
-}
-
-pub async fn probe_dictionary_params(
-    target: &Target,
-    args: &ScanArgs,
-    reflection_params: Arc<Mutex<Vec<Param>>>,
-    semaphore: Arc<Semaphore>,
-    pb: Option<ShimmerSpinner>,
-) {
-    let arc_target = Arc::new(target.clone());
-    let silence = args.silence;
-    let client = target.build_client_or_default();
-    // Taken before this stage pushes anything, so a collapse below can drop
-    // what the wordlist mined without touching what discovery already found.
-    let preexisting = snapshot_param_slots(&reflection_params).await;
-
-    // Resolve candidate parameter names (remote, file, or built-ins)
-    let mut params: Vec<String> = Vec::new();
-    let mut loaded = false;
-
-    if !args.remote_wordlists.is_empty() {
-        if let Err(e) = crate::payload::init_remote_wordlists(&args.remote_wordlists).await
-            && !silence
-        {
-            eprintln!("Error initializing remote wordlists: {}", e);
-        }
-        // Keyed by this scan's provider set: the cache is process-global, so a
-        // provider-less lookup in the server/MCP daemon could return whatever
-        // wordlist an earlier job with different providers had fetched.
-        if let Some(words) = crate::payload::get_remote_words_for(&args.remote_wordlists)
-            && !words.is_empty()
-        {
-            params = words.as_ref().clone();
-            loaded = true;
-        }
-    }
-
-    if !loaded && let Some(wordlist_path) = &args.mining_dict_word {
-        match crate::utils::fs::read_bounded(
-            std::path::Path::new(wordlist_path),
-            crate::utils::fs::MAX_FILE_READ_BYTES,
-            "parameter wordlist",
-        ) {
-            Ok(content) => {
-                params = content
-                    .lines()
-                    .map(|s| s.trim().to_string())
-                    .filter(|s| !s.is_empty())
-                    .collect();
-                loaded = true;
-            }
-            Err(e) => {
-                // Always surface on stderr — an unreadable
-                // `--mining-dict-word` is a user-supplied input error and
-                // silencing it (because server/MCP pass silence=true to
-                // analyze_parameters) means the operator can't tell why
-                // their custom dictionary did nothing. stderr never
-                // pollutes the stdout JSON/JSONL payload anyway.
-                eprintln!("Error reading wordlist file {}: {}", wordlist_path, e);
-                let _ = silence; // intentionally unused now
-                return;
-            }
-        }
-    }
-
-    if !loaded {
-        params = GF_PATTERNS_PARAMS.iter().map(ToString::to_string).collect();
-    }
-
-    // Sentinel pre-probe: 3 unique random param names. If every one reflects,
-    // the page echoes arbitrary input and the wordlist would just balloon
-    // into Stage 3-6 cost. Skip the wordlist and add a single "any" param
-    // instead — the params discovered before this stage are kept and still
-    // scanned. Skip when the wordlist is small enough that the pre-probe is
-    // more expensive than just running it.
-    if params.len() > SENTINEL_PROBE_COUNT * 5
-        && let Some(text) = pre_collapse_query_probe(&client, target).await
-    {
-        if !silence {
-            eprintln!(
-                "[mining-collapse] sentinel pre-probe collapsed Query mining: \
-                 every random param name reflected; adding single 'any' param"
-            );
-        }
-        collapse_mined_params(
-            &reflection_params,
-            &preexisting,
-            Location::Query,
-            Some(&text),
-        )
-        .await;
-        if let Some(ref pb) = pb {
-            pb.finish_and_clear();
-        }
-        return;
-    }
-
-    if let Some(ref pb) = pb {
-        pb.set_length(params.len() as u64);
-        pb.set_message("Mining dictionary parameters");
-    }
-
-    // EWMA adaptive stats shared across tasks
-    let stats = Arc::new(Mutex::new(MiningSampleStats::new()));
-
-    // Chunked processing bounds memory and allows early collapse exit. Tasks
-    // are spawned AND drained one chunk at a time (see the per-chunk join below)
-    // so the live task count never exceeds CHUNK_SIZE. Accumulating handles for
-    // the whole wordlist (joined only after every chunk) let a multi-million-line
-    // `--mining-dict-word` / `--remote-wordlists` spawn one task per entry up
-    // front — each holding cloned Url/String/Arc state — and blow resident
-    // memory into the tens of GB before any task was awaited.
-    const CHUNK_SIZE: usize = 500;
-    'outer: for param_chunk in params.chunks(CHUNK_SIZE) {
-        {
-            let st = stats.lock().await;
-            if st.collapsed {
-                break 'outer;
-            }
-        }
-        // Per-chunk handles, drained at the end of this chunk (see below).
-        let mut handles: Vec<tokio::task::JoinHandle<Option<Param>>> = Vec::new();
-        for param in param_chunk {
-            // Early collapse stop. (A second identical `collapsed` check used to
-            // follow immediately with no await in between — dead code, since
-            // this `break 'outer` already fires first when collapsed is set.)
-            {
-                let st = stats.lock().await;
-                if st.collapsed {
-                    break 'outer;
-                }
-            }
-            // Skip only when this *slot* was already discovered. Matching on
-            // the name alone let a param found at another wire location (a
-            // query `q`, a header `q`) suppress mining of the same name here,
-            // so the slot was never probed at all. See `param_slot_key`.
-            let exists = reflection_params
-                .lock()
-                .await
-                .iter()
-                .any(|p| p.name == *param && p.location == Location::Query);
-            if exists {
-                continue;
-            }
-
-            let mut url = target.url.clone();
-            url.query_pairs_mut()
-                .append_pair(param, crate::scanning::markers::bracketed_marker());
-
-            let client_clone = client.clone();
-
-            let data = target.data.clone();
-            let parsed_method = target.parse_method();
-            let target_clone = arc_target.clone();
-            let delay = target.delay;
-            let semaphore_clone = semaphore.clone();
-            let param_name = param.clone();
-            let pb_clone = pb.clone();
-            let stats_clone = stats.clone();
-
-            let handle = tokio::spawn(crate::with_job_scopes(
-                crate::JobScopes::capture(),
-                async move {
-                    let permit = semaphore_clone
-                        .acquire()
-                        .await
-                        .expect("acquire semaphore permit");
-                    let request = crate::utils::build_request(
-                        &client_clone,
-                        &target_clone,
-                        parsed_method,
-                        url,
-                        data.clone(),
-                    );
-
-                    crate::record_outbound_request().await;
-                    let resp = crate::utils::http::send_counted(request).await;
-
-                    let mut discovered: Option<Param> = None;
-                    if let Ok(r) = resp {
-                        // Skip server error responses (5xx) — debug error pages often
-                        // reflect query params in stack traces, causing false positives.
-                        let status = r.status();
-                        if status.is_server_error() {
-                            let mut st = stats_clone.lock().await;
-                            st.record_attempt();
-                            drop(permit);
-                            if delay > 0 {
-                                sleep(Duration::from_millis(delay)).await;
-                            }
-                            if let Some(ref pb) = pb_clone {
-                                pb.inc(1);
-                            }
-                            return discovered;
-                        }
-                        // Check for redirect reflection: if the response is a 3xx redirect,
-                        // the Location header may contain the reflected marker value.
-                        let is_redirect = status.is_redirection();
-                        let location_has_marker = if is_redirect {
-                            r.headers()
-                                .get("location")
-                                .and_then(|v| v.to_str().ok())
-                                .is_some_and(|loc| {
-                                    crate::scanning::markers::classify_probe_reflection(loc)
-                                        .detected()
-                                })
-                        } else {
-                            false
-                        };
-
-                        if location_has_marker {
-                            // Redirect context: marker reflected in Location header.
-                            let mut st = stats_clone.lock().await;
-                            st.record_attempt();
-                            st.record_reflection();
-                            if !st.collapsed {
-                                discovered = Some(Param {
-                                    injection_context: Some(
-                                        crate::parameter_analysis::InjectionContext::AttributeUrl(
-                                            None,
-                                        ),
-                                    ),
-                                    ..Param::new(
-                                        param_name.clone(),
-                                        crate::scanning::markers::bracketed_marker().to_string(),
-                                        crate::parameter_analysis::Location::Query,
-                                    )
-                                });
-                                if !silence {
-                                    eprintln!(
-                                        "Discovered parameter (redirect): {} (EWMA {:.2}, {}/{})",
-                                        param_name, st.ewma_ratio, st.reflections, st.attempts
-                                    );
-                                }
-                                if st.should_collapse() {
-                                    st.collapsed = true;
-                                    if !silence {
-                                        eprintln!(
-                                            "[mining-collapse] High reflection EWMA {:.2} after {} attempts ({} reflections)",
-                                            st.ewma_ratio, st.attempts, st.reflections
-                                        );
-                                    }
-                                }
-                            }
-                        } else if let Ok(text) = crate::utils::http::read_body(r).await {
-                            let mut st = stats_clone.lock().await;
-                            st.record_attempt();
-                            if crate::scanning::markers::classify_probe_reflection(&text).detected()
-                            {
-                                st.record_reflection();
-                                if !st.collapsed {
-                                    discovered = Some(
-                                        Param::new(
-                                            param_name.clone(),
-                                            crate::scanning::markers::bracketed_marker()
-                                                .to_string(),
-                                            crate::parameter_analysis::Location::Query,
-                                        )
-                                        .with_reflection_analysis(&text),
-                                    );
-                                    if !silence {
-                                        eprintln!(
-                                            "Discovered parameter: {} (EWMA {:.2}, {}/{})",
-                                            param_name, st.ewma_ratio, st.reflections, st.attempts
-                                        );
-                                    }
-                                    if st.should_collapse() {
-                                        st.collapsed = true;
-                                        if !silence {
-                                            eprintln!(
-                                                "[mining-collapse] High reflection EWMA {:.2} after {} attempts ({} reflections)",
-                                                st.ewma_ratio, st.attempts, st.reflections
-                                            );
-                                        }
-                                    }
-                                }
-                            } else {
-                                st.record_non_reflection();
-                            }
-                        }
-                    }
-
-                    if delay > 0 {
-                        sleep(Duration::from_millis(delay)).await;
-                    }
-                    drop(permit);
-                    if let Some(ref pb) = pb_clone {
-                        pb.inc(1);
-                    }
-                    discovered
-                },
-            ));
-
-            handles.push(handle);
-        }
-
-        // Drain THIS chunk's tasks (and flush discovered params) before the next
-        // chunk spawns, keeping the live task/handle count bounded by CHUNK_SIZE.
-        let mut batch: Vec<Param> = Vec::new();
-        for h in handles {
-            if let Ok(opt) = h.await
-                && let Some(p) = opt
-            {
-                batch.push(p);
-            }
-        }
-        if !batch.is_empty() {
-            let mut guard = reflection_params.lock().await;
-            guard.extend(batch);
-        }
-    } // end chunk loop
-
-    // Apply collapse post-processing once (instead of inside tasks mutating aggressively).
-    // Only the Query params *this stage mined* collapse — params discovered via
-    // other channels (Body, Header, Path, JsonBody, …) and the Query params
-    // Stage 1 discovery already confirmed are left alone.
-    let st_final = stats.lock().await;
-    if st_final.collapsed {
-        collapse_mined_params(&reflection_params, &preexisting, Location::Query, None).await;
-    }
-}
-
-pub async fn probe_body_params(
-    target: &Target,
-    args: &ScanArgs,
-    reflection_params: Arc<Mutex<Vec<Param>>>,
-    semaphore: Arc<Semaphore>,
-    pb: Option<ShimmerSpinner>,
-) {
-    let arc_target = Arc::new(target.clone());
-    let silence = args.silence;
-    let client = target.build_client_or_default();
-    let preexisting = snapshot_param_slots(&reflection_params).await;
-
-    if let Some(data) = &args.data {
-        // Assume form data for now (application/x-www-form-urlencoded)
-        let params: Vec<(String, String)> = form_urlencoded::parse(data.as_bytes())
-            .map(|(k, v)| (k.to_string(), v.to_string()))
-            .collect();
-
-        if let Some(ref pb) = pb {
-            pb.set_length(params.len() as u64);
-            pb.set_message("Mining body parameters");
-        }
-
-        // Adaptive EWMA stats shared across tasks
-        let stats = Arc::new(Mutex::new(MiningSampleStats::new()));
-
-        // Spawn tasks returning Option<Param> for batching
-        let mut handles: Vec<tokio::task::JoinHandle<Option<Param>>> = Vec::new();
-
-        for (param_name, _) in params {
-            // Early stop if collapsed
-            {
-                let st = stats.lock().await;
-                if st.collapsed {
-                    break;
-                }
-            }
-            // Skip already discovered params — but only at *this* wire slot.
-            // Keying on the name alone meant an ordinary
-            // `dalfox scan '…?q=x' -d 'q=y'` never mined the body `q` at all,
-            // because Stage 1 had already discovered the query `q`: a
-            // vulnerable body parameter was not just unreported, it was never
-            // probed. See `param_slot_key`.
-            let exists = reflection_params
-                .lock()
-                .await
-                .iter()
-                .any(|p| p.name == param_name && p.location == Location::Body);
-            if exists {
-                continue;
-            }
-
-            // Build mutated body with this param set to marker
-            let new_data = form_urlencoded::parse(data.as_bytes())
-                .map(|(k, v)| {
-                    if k == param_name {
-                        (k, crate::scanning::markers::bracketed_marker().to_string())
-                    } else {
-                        (k, v.to_string())
-                    }
-                })
-                .collect::<Vec<_>>();
-            let body = form_urlencoded::Serializer::new(String::new())
-                .extend_pairs(new_data)
-                .finish();
-
-            let client_clone = client.clone();
-            let url = target.url.clone();
-
-            let parsed_method = crate::scanning::url_inject::body_location_method(&target.method);
-            let target_clone = arc_target.clone();
-            let delay = target.delay;
-            let semaphore_clone = semaphore.clone();
-            let param_name_cloned = param_name.clone();
-            let pb_clone = pb.clone();
-            let stats_clone = stats.clone();
-
-            let handle = tokio::spawn(crate::with_job_scopes(
-                crate::JobScopes::capture(),
-                async move {
-                    let permit = semaphore_clone
-                        .acquire()
-                        .await
-                        .expect("acquire semaphore permit");
-                    let m = parsed_method;
-                    let base = crate::utils::build_body_request_base(
-                        &client_clone,
-                        &target_clone,
-                        m,
-                        url,
-                        Some(body),
-                    );
-                    let overrides = vec![(
-                        "Content-Type".to_string(),
-                        "application/x-www-form-urlencoded".to_string(),
-                    )];
-                    let request = crate::utils::apply_header_overrides(base, &overrides);
-
-                    crate::record_outbound_request().await;
-                    let resp = crate::utils::http::send_counted(request).await;
-
-                    let mut discovered: Option<Param> = None;
-                    if let Ok(r) = resp
-                        && let Ok(text) = crate::utils::http::read_body(r).await
-                    {
-                        let mut st = stats_clone.lock().await;
-                        st.record_attempt();
-                        if crate::scanning::markers::classify_probe_reflection(&text).detected() {
-                            st.record_reflection();
-                            if !st.collapsed {
-                                discovered = Some(
-                                    Param::new(
-                                        param_name_cloned.clone(),
-                                        crate::scanning::markers::bracketed_marker().to_string(),
-                                        Location::Body,
-                                    )
-                                    .with_reflection_analysis(&text),
-                                );
-                                if !silence {
-                                    eprintln!(
-                                        "Discovered body param: {} (EWMA {:.2}, {}/{})",
-                                        param_name_cloned,
-                                        st.ewma_ratio,
-                                        st.reflections,
-                                        st.attempts
-                                    );
-                                }
-                                if st.should_collapse() {
-                                    st.collapsed = true;
-                                    if !silence {
-                                        eprintln!(
-                                            "[mining-collapse] Body mining collapsed at EWMA {:.2} after {} attempts ({} reflections)",
-                                            st.ewma_ratio, st.attempts, st.reflections
-                                        );
-                                    }
-                                }
-                            }
-                        } else {
-                            st.record_non_reflection();
-                        }
-                    }
-
-                    if delay > 0 {
-                        sleep(Duration::from_millis(delay)).await;
-                    }
-                    drop(permit);
-                    if let Some(ref pb) = pb_clone {
-                        pb.inc(1);
-                    }
-                    discovered
-                },
-            ));
-
-            handles.push(handle);
-        }
-
-        // Batch collect discovered params
-        let mut batch: Vec<Param> = Vec::new();
-        for h in handles {
-            if let Ok(opt) = h.await
-                && let Some(p) = opt
-            {
-                batch.push(p);
-            }
-        }
-        if !batch.is_empty() {
-            let mut guard = reflection_params.lock().await;
-            guard.extend(batch);
-        }
-
-        // If collapsed after attempts, normalize the Body params this stage
-        // mined to a single 'any' param. Params discovered via other channels
-        // — and the Body params that were already known — are preserved.
-        let st_final = stats.lock().await;
-        if st_final.collapsed {
-            collapse_mined_params(&reflection_params, &preexisting, Location::Body, None).await;
-        }
-    }
-}
-
-pub async fn probe_response_id_params(
-    target: &Target,
-    args: &ScanArgs,
-    reflection_params: Arc<Mutex<Vec<Param>>>,
-    semaphore: Arc<Semaphore>,
-    pb: Option<ShimmerSpinner>,
-) {
-    let arc_target = Arc::new(target.clone());
-    let silence = args.silence;
-    let client = target.build_client_or_default();
-    let preexisting = snapshot_param_slots(&reflection_params).await;
-
-    // First, get the HTML to find input ids and names
-    let base_request = crate::utils::build_request(
-        &client,
-        target,
-        target.parse_method(),
-        target.url.clone(),
-        target.data.clone(),
-    );
-
-    crate::record_outbound_request().await;
-    let __resp = crate::utils::http::send_counted(base_request).await;
-    if let Ok(resp) = __resp
-        && !resp.status().is_server_error()
-        && let Ok(text) = crate::utils::http::read_body(resp).await
-    {
-        // Scope the scraper::Html (which is !Send) strictly to the owning
-        // block so the compiler can prove it never crosses any of the
-        // subsequent .await points. Extract owned String data, then drop
-        // the document before hitting the async code below.
-        let params_to_check: std::collections::HashSet<String> = {
-            let document = crate::utils::html::parse_document_bounded(&text);
-            let selector = selectors::input_with_id_or_name();
-            let mut set = std::collections::HashSet::new();
-            for element in document.select(selector) {
-                if let Some(id) = element.value().attr("id") {
-                    set.insert(id.to_string());
-                }
-                if let Some(name) = element.value().attr("name") {
-                    set.insert(name.to_string());
-                }
-            }
-            set
-        };
-
-        // Cap the DOM candidate set so a hostile/huge response body (up to the
-        // 16 MiB read_body cap) packed with distinct `id`/`name` attributes
-        // can't fan out into ~10^6 probe tasks + outbound requests. HashSet
-        // iteration order is arbitrary, which is fine for a safety ceiling:
-        // real pages stay far under it. Probing then consumes this Vec one
-        // owned name at a time exactly as it did the set.
-        let (params_to_check, capped_from) = cap_dom_params(params_to_check.into_iter().collect());
-        if let Some(original) = capped_from
-            && !silence
-        {
-            eprintln!(
-                "[mining] DOM candidate params capped to {} (from {}); reduce reflected fields or use --skip-mining",
-                MAX_DOM_MINING_PARAMS, original
-            );
-        }
-
-        // Sentinel pre-probe — same rationale as Query mining: a
-        // reflect-everything page would mark every DOM-extracted name as
-        // reflected and balloon downstream cost. Threshold matches Query
-        // mining: only run when the candidate set exceeds the pre-probe
-        // ceiling.
-        if params_to_check.len() > SENTINEL_PROBE_COUNT * 5
-            && let Some(text) = pre_collapse_query_probe(&client, target).await
-        {
-            if !silence {
-                eprintln!(
-                    "[mining-collapse] sentinel pre-probe collapsed DOM mining: \
-                     every random param name reflected; adding single 'any' param"
-                );
-            }
-            collapse_mined_params(
-                &reflection_params,
-                &preexisting,
-                Location::Query,
-                Some(&text),
-            )
-            .await;
-            if let Some(ref pb) = pb {
-                pb.finish_and_clear();
-            }
-            return;
-        }
-
-        if let Some(ref pb) = pb {
-            pb.set_length(params_to_check.len() as u64);
-            pb.set_message("Mining DOM parameters");
-        }
-
-        // Spawn tasks returning Option<Param> for batched collection
-        let mut handles: Vec<tokio::task::JoinHandle<Option<Param>>> = Vec::new();
-        let stats = Arc::new(Mutex::new(MiningSampleStats::new()));
-
-        // Check each param for reflection
-        for param in params_to_check {
-            {
-                let st = stats.lock().await;
-                if st.collapsed {
-                    break;
-                }
-            }
-            // Slot-scoped, not name-scoped: see `param_slot_key`.
-            let existing = reflection_params
-                .lock()
-                .await
-                .iter()
-                .any(|p| p.name == param && p.location == Location::Query);
-            if existing {
-                continue;
-            }
-            let mut url = target.url.clone();
-            url.query_pairs_mut()
-                .append_pair(&param, crate::scanning::markers::bracketed_marker());
-            let client_clone = client.clone();
-
-            let data = target.data.clone();
-            let parsed_method = target.parse_method();
-            let target_clone = arc_target.clone();
-            let delay = target.delay;
-            let semaphore_clone = semaphore.clone();
-            let param = param.clone();
-            let pb_clone = pb.clone();
-            let stats_clone = stats.clone();
-
-            let handle = tokio::spawn(crate::with_job_scopes(
-                crate::JobScopes::capture(),
-                async move {
-                    let permit = semaphore_clone
-                        .acquire()
-                        .await
-                        .expect("acquire semaphore permit");
-                    let m = parsed_method;
-                    let request = crate::utils::build_request(
-                        &client_clone,
-                        &target_clone,
-                        m,
-                        url,
-                        data.clone(),
-                    );
-                    // Prepare optional discovered Param container for batched return
-                    let mut discovered: Option<Param> = None;
-                    crate::record_outbound_request().await;
-                    let __resp = crate::utils::http::send_counted(request).await;
-                    if let Ok(resp) = __resp {
-                        // Skip 5xx error responses — debug pages often reflect params
-                        if resp.status().is_server_error() {
-                            let mut st = stats_clone.lock().await;
-                            st.record_attempt();
-                            drop(permit);
-                            if delay > 0 {
-                                sleep(Duration::from_millis(delay)).await;
-                            }
-                            if let Some(ref pb) = pb_clone {
-                                pb.inc(1);
-                            }
-                            return discovered;
-                        }
-                        if let Ok(text) = crate::utils::http::read_body(resp).await {
-                            let mut st = stats_clone.lock().await;
-                            st.record_attempt();
-                            if crate::scanning::markers::classify_probe_reflection(&text).detected()
-                            {
-                                st.record_reflection();
-                                if !st.collapsed {
-                                    // Store discovered Param for return (batched later)
-                                    discovered = Some(
-                                        Param::new(
-                                            param.clone(),
-                                            crate::scanning::markers::bracketed_marker()
-                                                .to_string(),
-                                            crate::parameter_analysis::Location::Query,
-                                        )
-                                        .with_reflection_analysis(&text),
-                                    );
-                                    if !silence {
-                                        eprintln!(
-                                            "Discovered DOM param: {} (EWMA {:.2}, {}/{})",
-                                            param, st.ewma_ratio, st.reflections, st.attempts
-                                        );
-                                    }
-                                    if st.should_collapse() {
-                                        st.collapsed = true;
-                                        if !silence {
-                                            eprintln!(
-                                                "[mining-collapse] DOM mining collapsed at EWMA {:.2} after {} attempts ({} reflections)",
-                                                st.ewma_ratio, st.attempts, st.reflections
-                                            );
-                                        }
-                                    }
-                                }
-                            } else {
-                                st.record_non_reflection();
-                            }
-                        }
-                    }
-                    if delay > 0 {
-                        sleep(Duration::from_millis(delay)).await;
-                    }
-                    drop(permit);
-                    if let Some(ref pb) = pb_clone {
-                        pb.inc(1);
-                    }
-                    // Return discovered Param (if any) for batch processing
-                    discovered
-                },
-            ));
-            handles.push(handle);
-        }
-
-        // Batch collect discovered DOM params
-        let mut batch: Vec<Param> = Vec::new();
-        for handle in handles {
-            if let Ok(opt) = handle.await
-                && let Some(p) = opt
-            {
-                batch.push(p);
-            }
-        }
-        if !batch.is_empty() {
-            let mut guard = reflection_params.lock().await;
-            guard.extend(batch);
-        }
-        // Collapse post-processing (single 'any' param) if adaptive stats
-        // triggered it. Only the Query params this stage mined are folded in;
-        // params discovered via other channels — and the ones discovery already
-        // confirmed — survive.
-        let st_final = stats.lock().await;
-        if st_final.collapsed {
-            collapse_mined_params(&reflection_params, &preexisting, Location::Query, None).await;
-        }
-    }
-}
-
-pub async fn probe_json_body_params(
-    target: &Target,
-    args: &ScanArgs,
-    reflection_params: Arc<Mutex<Vec<Param>>>,
-    semaphore: Arc<Semaphore>,
-    pb: Option<ShimmerSpinner>,
-) {
-    let arc_target = Arc::new(target.clone());
-    let silence = args.silence;
-    let client = target.build_client_or_default();
-    let preexisting = snapshot_param_slots(&reflection_params).await;
-
-    // Detect JSON body from args.data; only proceed if it's a JSON object
-    let base_json: serde_json::Value = match &args.data {
-        Some(d) => match serde_json::from_str::<serde_json::Value>(d) {
-            Ok(v) => v,
-            Err(_) => return, // not JSON
-        },
-        None => return,
-    };
-    if !base_json.is_object() {
-        return;
-    }
-    // A GraphQL request is JSON too, but its injection points live inside the
-    // `variables` object (handled by `probe_graphql_params`, which rebuilds the
-    // whole request per variable). Mining the top-level `query`/`variables`
-    // keys here would replace the `variables` object with a bare marker string
-    // and garble the request, so hand GraphQL bodies off entirely.
-    if let Some(data) = &args.data
-        && !crate::encoding::pipeline::infer_graphql_variable_fields(data).is_empty()
-    {
-        return;
-    }
-
-    // Collect top-level keys to mutate
-    let Some(obj) = base_json.as_object() else {
-        return;
-    };
-    let keys: Vec<String> = obj.keys().cloned().collect();
-
-    if let Some(ref pb) = pb {
-        pb.set_length(keys.len() as u64);
-        pb.set_message("Mining JSON body parameters");
-    }
-
-    // Adaptive EWMA stats shared across tasks
-    let stats = Arc::new(Mutex::new(MiningSampleStats::new()));
-
-    // Spawn tasks returning Option<Param> for batching
-    let mut handles: Vec<tokio::task::JoinHandle<Option<Param>>> = Vec::new();
-
-    for param_name in keys {
-        {
-            // Early collapse stop
-            let st = stats.lock().await;
-            if st.collapsed {
-                break;
-            }
-        }
-        // Slot-scoped, not name-scoped: a query or form `q` must not suppress
-        // mining of the JSON body's `q`. See `param_slot_key`.
-        let exists = reflection_params
-            .lock()
-            .await
-            .iter()
-            .any(|p| p.name == param_name && p.location == Location::JsonBody);
-        if exists {
-            continue;
-        }
-
-        let client_clone = client.clone();
-        let url = target.url.clone();
-
-        let parsed_method = crate::scanning::url_inject::body_location_method(&target.method);
-        let target_clone = arc_target.clone();
-        let delay = target.delay;
-        let semaphore_clone = semaphore.clone();
-        let param_name_cloned = param_name.clone();
-        let pb_clone = pb.clone();
-        let stats_clone = stats.clone();
-        let base_json_clone = base_json.clone();
-
-        let handle = tokio::spawn(crate::with_job_scopes(
-            crate::JobScopes::capture(),
-            async move {
-                let permit = semaphore_clone
-                    .acquire()
-                    .await
-                    .expect("acquire semaphore permit");
-
-                // Build mutated JSON with this key set to marker
-                let mut root = base_json_clone;
-                if let Some(map) = root.as_object_mut() {
-                    map.insert(
-                        param_name_cloned.clone(),
-                        serde_json::Value::String(
-                            crate::scanning::markers::bracketed_marker().to_string(),
-                        ),
-                    );
-                } else {
-                    let mut map = serde_json::Map::new();
-                    map.insert(
-                        param_name_cloned.clone(),
-                        serde_json::Value::String(
-                            crate::scanning::markers::bracketed_marker().to_string(),
-                        ),
-                    );
-                    root = serde_json::Value::Object(map);
-                }
-                let body = serde_json::to_string(&root).unwrap_or_else(|_| {
-                    format!(
-                        "{{\"{}\":\"{}\"}}",
-                        param_name_cloned,
-                        crate::scanning::markers::bracketed_marker()
-                    )
-                });
-
-                let base = crate::utils::build_body_request_base(
-                    &client_clone,
-                    &target_clone,
-                    parsed_method,
-                    url,
-                    Some(body),
-                );
-                let overrides = vec![("Content-Type".to_string(), "application/json".to_string())];
-                let request = crate::utils::apply_header_overrides(base, &overrides);
-
-                crate::record_outbound_request().await;
-                let resp = crate::utils::http::send_counted(request).await;
-
-                let mut discovered: Option<Param> = None;
-                if let Ok(r) = resp
-                    && let Ok(text) = crate::utils::http::read_body(r).await
-                {
-                    let mut st = stats_clone.lock().await;
-                    st.record_attempt();
-                    if crate::scanning::markers::classify_probe_reflection(&text).detected() {
-                        st.record_reflection();
-                        if !st.collapsed {
-                            discovered = Some(
-                                Param::new(
-                                    param_name_cloned.clone(),
-                                    crate::scanning::markers::bracketed_marker().to_string(),
-                                    Location::JsonBody,
-                                )
-                                .with_reflection_analysis(&text),
-                            );
-                            if !silence {
-                                eprintln!(
-                                    "Discovered JSON body param: {} (EWMA {:.2}, {}/{})",
-                                    param_name_cloned, st.ewma_ratio, st.reflections, st.attempts
-                                );
-                            }
-                            if st.should_collapse() {
-                                st.collapsed = true;
-                                if !silence {
-                                    eprintln!(
-                                        "[mining-collapse] JSON mining collapsed at EWMA {:.2} after {} attempts ({} reflections)",
-                                        st.ewma_ratio, st.attempts, st.reflections
-                                    );
-                                }
-                            }
-                        }
-                    } else {
-                        st.record_non_reflection();
-                    }
-                }
-
-                if delay > 0 {
-                    sleep(Duration::from_millis(delay)).await;
-                }
-                drop(permit);
-                if let Some(ref pb) = pb_clone {
-                    pb.inc(1);
-                }
-                discovered
-            },
-        ));
-
-        handles.push(handle);
-    }
-
-    // Batch collect discovered params
-    let mut batch: Vec<Param> = Vec::new();
-    for h in handles {
-        if let Ok(opt) = h.await
-            && let Some(p) = opt
-        {
-            batch.push(p);
-        }
-    }
-    if !batch.is_empty() {
-        let mut guard = reflection_params.lock().await;
-        guard.extend(batch);
-    }
-
-    // Collapse normalization to single 'any' JSON param if triggered. Only the
-    // JsonBody params this stage mined fold in; everything else is preserved.
-    let st_final = stats.lock().await;
-    if st_final.collapsed {
-        collapse_mined_params(&reflection_params, &preexisting, Location::JsonBody, None).await;
-    }
-}
-
-/// Seed GraphQL injection points from an `-d` GraphQL-over-JSON body.
-///
-/// Detects the GraphQL request shape (see
-/// [`crate::encoding::pipeline::infer_graphql_variable_fields`]) and probes each
-/// string value inside the `variables` object by rebuilding the whole request
-/// with a marker in that one slot. A field is seeded as a
-/// [`Location::GraphqlBody`] param — carrying the whole-request rebuild
-/// pipeline — only when the marker reflects, so an ordinary JSON API can never
-/// produce GraphQL params.
-///
-/// No-op unless `-d` parses as a GraphQL request, so REST/JSON/GET targets are
-/// unaffected. Bounded by the pipeline walker's leaf cap (≤32 fields).
-pub async fn probe_graphql_params(
-    target: &Target,
-    args: &ScanArgs,
-    reflection_params: Arc<Mutex<Vec<Param>>>,
-    semaphore: Arc<Semaphore>,
-    pb: Option<ShimmerSpinner>,
-) {
-    let Some(data) = &args.data else {
-        return;
-    };
-    let fields = crate::encoding::pipeline::infer_graphql_variable_fields(data);
-    if fields.is_empty() {
-        return;
-    }
-
-    if let Some(ref pb) = pb {
-        pb.set_length(fields.len() as u64);
-        pb.set_message("Probing GraphQL variables");
-    }
-
-    let arc_target = Arc::new(target.clone());
-    let client = target.build_client_or_default();
-    let silence = args.silence;
-    let marker = crate::scanning::markers::bracketed_marker();
-
-    let mut handles: Vec<tokio::task::JoinHandle<Option<Param>>> = Vec::new();
-    for nf in fields {
-        // Display name: `variables.<path>` (e.g. `variables.input.title`).
-        let name = nf.path.join(".");
-        // Slot-scoped skip: only suppress a GraphQL field already seeded here.
-        let exists = reflection_params
-            .lock()
-            .await
-            .iter()
-            .any(|p| p.name == name && p.location == Location::GraphqlBody);
-        if exists {
-            continue;
-        }
-
-        let client_clone = client.clone();
-        let url = target.url.clone();
-        let target_clone = arc_target.clone();
-        let semaphore_clone = semaphore.clone();
-        let delay = target.delay;
-        let pb_clone = pb.clone();
-        let pipeline = nf.pipeline.clone();
-        let original_value = nf.original_value.clone();
-        let name_for_task = name.clone();
-
-        let handle = tokio::spawn(crate::with_job_scopes(
-            crate::JobScopes::capture(),
-            async move {
-                let permit = semaphore_clone
-                    .acquire()
-                    .await
-                    .expect("acquire semaphore permit");
-
-                // Rebuild the full request with the marker in this one variable.
-                let body = match pipeline.apply(marker) {
-                    Ok(b) => b,
-                    Err(_) => {
-                        drop(permit);
-                        return None;
-                    }
-                };
-
-                let base = crate::utils::build_body_request_base(
-                    &client_clone,
-                    &target_clone,
-                    crate::scanning::url_inject::body_location_method(&target_clone.method),
-                    url,
-                    Some(body),
-                );
-                let request = crate::utils::apply_header_overrides(
-                    base,
-                    &[("Content-Type".to_string(), "application/json".to_string())],
-                );
-
-                crate::record_outbound_request().await;
-                let mut discovered: Option<Param> = None;
-                if let Ok(r) = crate::utils::http::send_counted(request).await
-                    && let Ok(text) = crate::utils::http::read_body(r).await
-                    && crate::scanning::markers::classify_probe_reflection(&text).detected()
-                {
-                    if !silence {
-                        eprintln!("Discovered GraphQL variable: {}", name_for_task);
-                    }
-                    discovered = Some(
-                        Param {
-                            pre_encoding_pipeline: Some(pipeline),
-                            ..Param::new(name_for_task, original_value, Location::GraphqlBody)
-                        }
-                        .with_reflection_analysis(&text),
-                    );
-                }
-
-                if delay > 0 {
-                    sleep(Duration::from_millis(delay)).await;
-                }
-                drop(permit);
-                if let Some(ref pb) = pb_clone {
-                    pb.inc(1);
-                }
-                discovered
-            },
-        ));
-        handles.push(handle);
-    }
-
-    let mut batch: Vec<Param> = Vec::new();
-    for h in handles {
-        if let Ok(Some(p)) = h.await {
-            batch.push(p);
-        }
-    }
-    if !batch.is_empty() {
-        reflection_params.lock().await.extend(batch);
-    }
-}
-
-/// Whether the target's request body should be treated as XML for injection:
-/// its declared `Content-Type` is an XML-family type, or the body opens with an
-/// `<?xml` prolog. Deliberately strict so JSON / form / HTML bodies are not
-/// mistaken for XML.
-fn request_is_xml(target: &Target, data: &str) -> bool {
-    let ct_is_xml = target
-        .headers
-        .iter()
-        .find(|(k, _)| k.eq_ignore_ascii_case("content-type"))
-        .map(|(_, v)| {
-            matches!(
-                crate::utils::content_type_primary(v).as_deref(),
-                Some("text/xml")
-                    | Some("application/xml")
-                    | Some("application/soap+xml")
-                    | Some("application/xhtml+xml")
-            )
-        })
-        .unwrap_or(false);
-    ct_is_xml || data.trim_start().starts_with("<?xml")
-}
-
-/// Seed XML / SOAP injection points from an `-d` XML body.
-///
-/// Runs only when the request body is XML (see [`request_is_xml`]). Tokenizes
-/// the document into element text nodes and attribute values (see
-/// [`crate::parameter_analysis::xml_inject::xml_injection_points`]) and probes
-/// each by splicing a marker into that one leaf — the rest of the document
-/// survives byte-for-byte. A leaf is seeded as a [`Location::XmlBody`] param,
-/// carrying the exact-byte-range `Splice` pipeline, only when the marker
-/// reflects, so a benign XML body that echoes nothing produces no params.
-///
-/// No-op on JSON / form / GET targets. Bounded by the tokenizer's point cap.
-pub async fn probe_xml_body_params(
-    target: &Target,
-    args: &ScanArgs,
-    reflection_params: Arc<Mutex<Vec<Param>>>,
-    semaphore: Arc<Semaphore>,
-    pb: Option<ShimmerSpinner>,
-) {
-    let Some(data) = &args.data else {
-        return;
-    };
-    if !request_is_xml(target, data) {
-        return;
-    }
-    let points = crate::parameter_analysis::xml_inject::xml_injection_points(data);
-    if points.is_empty() {
-        return;
-    }
-
-    if let Some(ref pb) = pb {
-        pb.set_length(points.len() as u64);
-        pb.set_message("Probing XML body");
-    }
-
-    let arc_target = Arc::new(target.clone());
-    let client = target.build_client_or_default();
-    let silence = args.silence;
-    let marker = crate::scanning::markers::bracketed_marker();
-    let content_type = crate::scanning::url_inject::xml_request_content_type(target);
-
-    let mut handles: Vec<tokio::task::JoinHandle<Option<Param>>> = Vec::new();
-    for point in points {
-        let name = point.name.clone();
-        let exists = reflection_params
-            .lock()
-            .await
-            .iter()
-            .any(|p| p.name == name && p.location == Location::XmlBody);
-        if exists {
-            continue;
-        }
-
-        // Exact-byte-range splice: prefix + payload + suffix rebuilds the whole
-        // document with only this leaf replaced.
-        let pipeline = crate::encoding::pipeline::EncodingPipeline::new(vec![
-            crate::encoding::pipeline::EncodingStep::Splice {
-                prefix: data[..point.start].to_string(),
-                suffix: data[point.end..].to_string(),
-            },
-        ]);
-
-        let client_clone = client.clone();
-        let url = target.url.clone();
-        let target_clone = arc_target.clone();
-        let semaphore_clone = semaphore.clone();
-        let delay = target.delay;
-        let pb_clone = pb.clone();
-        let ct = content_type.clone();
-        let original_value = point.value.clone();
-        let name_for_task = name.clone();
-        let pipeline_for_task = pipeline.clone();
-
-        let handle = tokio::spawn(crate::with_job_scopes(
-            crate::JobScopes::capture(),
-            async move {
-                let permit = semaphore_clone
-                    .acquire()
-                    .await
-                    .expect("acquire semaphore permit");
-
-                let body = match pipeline_for_task.apply(marker) {
-                    Ok(b) => b,
-                    Err(_) => {
-                        drop(permit);
-                        return None;
-                    }
-                };
-
-                let base = crate::utils::build_body_request_base(
-                    &client_clone,
-                    &target_clone,
-                    crate::scanning::url_inject::body_location_method(&target_clone.method),
-                    url,
-                    Some(body),
-                );
-                let request =
-                    crate::utils::apply_header_overrides(base, &[("Content-Type".to_string(), ct)]);
-
-                crate::record_outbound_request().await;
-                let mut discovered: Option<Param> = None;
-                if let Ok(r) = crate::utils::http::send_counted(request).await
-                    && let Ok(text) = crate::utils::http::read_body(r).await
-                    && crate::scanning::markers::classify_probe_reflection(&text).detected()
-                {
-                    if !silence {
-                        eprintln!("Discovered XML injection point: {}", name_for_task);
-                    }
-                    discovered = Some(
-                        Param {
-                            pre_encoding_pipeline: Some(pipeline_for_task),
-                            ..Param::new(name_for_task, original_value, Location::XmlBody)
-                        }
-                        .with_reflection_analysis(&text),
-                    );
-                }
-
-                if delay > 0 {
-                    sleep(Duration::from_millis(delay)).await;
-                }
-                drop(permit);
-                if let Some(ref pb) = pb_clone {
-                    pb.inc(1);
-                }
-                discovered
-            },
-        ));
-        handles.push(handle);
-    }
-
-    let mut batch: Vec<Param> = Vec::new();
-    for h in handles {
-        if let Ok(Some(p)) = h.await {
-            batch.push(p);
-        }
-    }
-    if !batch.is_empty() {
-        reflection_params.lock().await.extend(batch);
-    }
-}
-
-/// Seed multipart form-field params the operator named via `-p name:multipart`.
-///
-/// Like [`probe_body_params`], these come from explicit `-d` input rather than
-/// discovery — without this, `MultipartBody` params were only ever seeded from
-/// HTML `<form enctype=multipart/form-data>` discovery, so `-p file:multipart`
-/// (a known multipart sink) had no entry point and was silently never tested.
-/// Sends a real `multipart/form-data` probe and seeds the field as a
-/// `MultipartBody` param when the marker reflects. No-op without both `-d` and
-/// at least one `-p :multipart` spec.
-pub async fn probe_multipart_params(
-    target: &Target,
-    args: &ScanArgs,
-    reflection_params: Arc<Mutex<Vec<Param>>>,
-    semaphore: Arc<Semaphore>,
-    pb: Option<ShimmerSpinner>,
-) {
-    let Some(data) = &args.data else {
-        return;
-    };
-    let wanted =
-        crate::parameter_analysis::discovery::explicit_param_names(&args.param, "multipart");
-    if wanted.is_empty() {
-        return;
-    }
-
-    if let Some(ref pb) = pb {
-        pb.set_length(wanted.len() as u64);
-        pb.set_message("Probing multipart fields");
-    }
-
-    let pairs: Vec<(String, String)> = form_urlencoded::parse(data.as_bytes())
-        .map(|(k, v)| (k.to_string(), v.to_string()))
-        .collect();
-    let client = target.build_client_or_default();
-    let marker = crate::scanning::markers::bracketed_marker();
-    let silence = args.silence;
-
-    let mut handles: Vec<tokio::task::JoinHandle<Option<Param>>> = Vec::new();
-    for field in wanted {
-        // Skip only if this name is already registered *as a multipart field*.
-        // A same-named body/query param (e.g. `probe_body_params` seeding
-        // `file` from the same `-d`) must not block the multipart slot —
-        // `-p file:multipart` filters by location, so the body entry would be
-        // dropped and we'd be left with nothing.
-        let exists = reflection_params
-            .lock()
-            .await
-            .iter()
-            .any(|p| p.name == field && p.location == Location::MultipartBody);
-        if exists {
-            continue;
-        }
-
-        let client_clone = client.clone();
-        let url = target.url.clone();
-        let target_clone = Arc::new(target.clone());
-        let semaphore_clone = semaphore.clone();
-        let pairs_clone = pairs.clone();
-        let field_name = field.clone();
-
-        let handle = tokio::spawn(crate::with_job_scopes(
-            crate::JobScopes::capture(),
-            async move {
-                let permit = semaphore_clone
-                    .acquire()
-                    .await
-                    .expect("acquire semaphore permit");
-                let mut form = reqwest::multipart::Form::new();
-                let mut placed = false;
-                for (k, v) in &pairs_clone {
-                    if *k == field_name {
-                        form = form.text(k.clone(), marker.to_string());
-                        placed = true;
-                    } else {
-                        form = form.text(k.clone(), v.clone());
-                    }
-                }
-                if !placed {
-                    form = form.text(field_name.clone(), marker.to_string());
-                }
-
-                let method =
-                    crate::scanning::url_inject::body_location_method(&target_clone.method);
-                let request = crate::utils::build_body_request_base(
-                    &client_clone,
-                    &target_clone,
-                    method,
-                    url,
-                    None,
-                )
-                .multipart(form);
-                crate::record_outbound_request().await;
-
-                let mut discovered: Option<Param> = None;
-                if let Ok(r) = crate::utils::http::send_counted(request).await
-                    && let Ok(text) = crate::utils::http::read_body(r).await
-                    && crate::scanning::markers::classify_probe_reflection(&text).detected()
-                {
-                    if !silence {
-                        eprintln!("Discovered multipart field: {}", field_name);
-                    }
-                    discovered = Some(
-                        Param::new(field_name, marker.to_string(), Location::MultipartBody)
-                            .with_reflection_analysis(&text),
-                    );
-                }
-                drop(permit);
-                discovered
-            },
-        ));
-        handles.push(handle);
-    }
-
-    let mut batch: Vec<Param> = Vec::new();
-    for handle in handles {
-        if let Ok(Some(p)) = handle.await {
-            batch.push(p);
-        }
-    }
-    if !batch.is_empty() {
-        reflection_params.lock().await.extend(batch);
-    }
-}
-
-pub async fn mine_parameters(
-    target: &mut Target,
-    args: &ScanArgs,
-    reflection_params: Arc<Mutex<Vec<Param>>>,
-    semaphore: Arc<Semaphore>,
-    pb: Option<ShimmerSpinner>,
-) {
-    // Body/JSON parameters supplied via `-d` are explicit user input, not
-    // discovery. Seed them independent of the mining flags: query params are
-    // seeded from the URL during the discovery stage, but body params have no
-    // other entry point, so gating them behind mining drops the entire
-    // POST/JSON body surface — even when `-p name:body` is given and even
-    // under `--skip-mining` / `--skip-mining-dict`. (Both probes are no-ops
-    // when `args.data` is None, so GET targets are unaffected.)
-    probe_body_params(
-        target,
-        args,
-        reflection_params.clone(),
-        semaphore.clone(),
-        pb.clone(),
-    )
-    .await;
-    probe_json_body_params(
-        target,
-        args,
-        reflection_params.clone(),
-        semaphore.clone(),
-        pb.clone(),
-    )
-    .await;
-    // GraphQL variables: own path for GraphQL-over-JSON bodies (no-op on plain
-    // JSON, which `probe_json_body_params` already handled).
-    probe_graphql_params(
-        target,
-        args,
-        reflection_params.clone(),
-        semaphore.clone(),
-        pb.clone(),
-    )
-    .await;
-    // XML / SOAP element text + attribute values (no-op on non-XML bodies).
-    probe_xml_body_params(
-        target,
-        args,
-        reflection_params.clone(),
-        semaphore.clone(),
-        pb.clone(),
-    )
-    .await;
-    probe_multipart_params(
-        target,
-        args,
-        reflection_params.clone(),
-        semaphore.clone(),
-        pb.clone(),
-    )
-    .await;
-
-    // Mining proper: discover parameters the user did NOT name.
-    if !args.skip_mining {
-        if !args.skip_mining_dict {
-            probe_dictionary_params(
-                target,
-                args,
-                reflection_params.clone(),
-                semaphore.clone(),
-                pb.clone(),
-            )
-            .await;
-        }
-        if !args.skip_mining_dom {
-            probe_response_id_params(
-                target,
-                args,
-                reflection_params.clone(),
-                semaphore.clone(),
-                pb.clone(),
-            )
-            .await;
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests;
